Move tests related to cocotb.test into their own module
